move files according to PR remarks
