+++ src/main/java/org/gridsuite/actions/server/ContingencyListService.java
@@ -18,6 +18,12 @@ import com.powsybl.network.store.client.PreloadingStrategy;
 import com.powsybl.network.store.iidm.impl.NetworkFactoryImpl;
 import org.gridsuite.actions.ContingencyListEvaluator;
 import org.gridsuite.actions.dto.*;
+import org.gridsuite.actions.dto.contingency.FilterBasedContingencyList;
+import org.gridsuite.actions.dto.contingency.IdBasedContingencyList;
+import org.gridsuite.actions.dto.contingency.PersistentContingencyList;
+import org.gridsuite.actions.dto.evaluation.ContingencyIdsByGroup;
+import org.gridsuite.actions.dto.evaluation.ContingencyInfos;
+import org.gridsuite.actions.dto.evaluation.ContingencyListExportResult;
 import org.gridsuite.actions.server.entities.*;
 import org.gridsuite.actions.server.repositories.FilterBasedContingencyListRepository;
 import org.gridsuite.actions.server.repositories.IdBasedContingencyListRepository;
